feat(split): split video feature working but split video slow in performance

+++ fileUtils.js
@@ -29,11 +29,16 @@ async function getFileSize(file) {
 	}
 }
 
-function deleteFile(file) {
-	return new Promise((resolve, reject) => {
-		fs.unlink(file);
-		console.log(`Deleted file: ${file}`);
-	});
+async function deleteFile(file) {
+	try {
+		if (file) {
+			await fs.unlink(file);
+			console.log(`Deleted file: ${file}`);
+		}
+	}
+	catch (error) {
+		console.error(error);
+	}
 }
 
 function getNameWithoutFileFormat(fileName) {
